use common dependency specifier for debian
