+++ syft/pkg/cataloger/debian/cataloger.go
@@ -15,5 +15,5 @@ func NewDBCataloger() pkg.Cataloger {
 		// note: these globs have been intentionally split up in order to improve search performance,
 		// please do NOT combine into: "**/var/lib/dpkg/{status,status.d/*}"
 		WithParserByGlobs(parseDpkgDB, "**/var/lib/dpkg/status", "**/var/lib/dpkg/status.d/*", "**/lib/opkg/info/*.control", "**/lib/opkg/status").
-		WithProcessors(dependency.Processor(newDBProsumer()))
+		WithProcessors(dependency.Processor(dbEntryDependencySpecifier))
 }
