+++ syft/pkg/cataloger/debian/parse_dpkg_db_test.go
@@ -2,7 +2,6 @@ package debian
 
 import (
 	"bufio"
-	"context"
 	"errors"
 	"fmt"
 	"os"
@@ -429,99 +428,6 @@ func Test_handleNewKeyValue(t *testing.T) {
 
 			assert.Equalf(t, tt.wantKey, gotKey, "handleNewKeyValue(%v)", tt.line)
 			assert.Equalf(t, tt.wantVal, gotVal, "handleNewKeyValue(%v)", tt.line)
-		})
-	}
-}
-
-func Test_stripVersionSpecifier(t *testing.T) {
-
-	tests := []struct {
-		name  string
-		input string
-		want  string
-	}{
-		{
-			name:  "package name only",
-			input: "test",
-			want:  "test",
-		},
-		{
-			name:  "with version",
-			input: "test (1.2.3)",
-			want:  "test",
-		},
-		{
-			name:  "multiple packages",
-			input: "test | other",
-			want:  "test | other",
-		},
-		{
-			name:  "with architecture specifiers",
-			input: "test [amd64 i386]",
-			want:  "test",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, stripVersionSpecifier(tt.input))
-		})
-	}
-}
-
-func Test_associateRelationships(t *testing.T) {
-	tests := []struct {
-		name              string
-		fixture           string
-		wantRelationships map[string][]string
-	}{
-		{
-			name:    "relationships for coreutils",
-			fixture: "test-fixtures/var/lib/dpkg/status.d/coreutils-relationships",
-			wantRelationships: map[string][]string{
-				"coreutils":    {"libacl1", "libattr1", "libc6", "libgmp10", "libselinux1"},
-				"libacl1":      {"libc6"},
-				"libattr1":     {"libc6"},
-				"libc6":        {"libgcc-s1"},
-				"libgcc-s1":    {"gcc-12-base", "libc6"},
-				"libgmp10":     {"libc6"},
-				"libpcre2-8-0": {"libc6"},
-				"libselinux1":  {"libc6", "libpcre2-8-0"},
-			},
-		},
-		{
-			name:    "relationships from dpkg example docs",
-			fixture: "test-fixtures/var/lib/dpkg/status.d/doc-examples",
-			wantRelationships: map[string][]string{
-				"made-up-package-1": {"gnumach-dev", "hurd-dev", "kernel-headers-2.2.10"},
-				"made-up-package-2": {"liblua5.1-dev", "libluajit5.1-dev"},
-				"made-up-package-3": {"bar", "foo"},
-				// note that the "made-up-package-4" depends on "made-up-package-5" but not via the direct
-				// package name, but through the "provides" virtual package name "virtual-package-5".
-				"made-up-package-4": {"made-up-package-5"},
-				// note that though there is a "default-mta | mail-transport-agent | not-installed"
-				// dependency choice we raise up the packages that are installed for every choice.
-				// In this case that means that "default-mta" and "mail-transport-agent".
-				"mutt": {"default-mta", "libc6", "mail-transport-agent"},
-			},
-		},
-		{
-			name:    "relationships for libpam-runtime",
-			fixture: "test-fixtures/var/lib/dpkg/status.d/libpam-runtime",
-			wantRelationships: map[string][]string{
-				"libpam-runtime": {"cdebconf", "debconf-2.0", "debconf1", "debconf2", "libpam-modules"},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			pkgs, relationships, err := NewDBCataloger().Catalog(context.Background(), file.NewMockResolverForPaths(tt.fixture))
-			require.NotEmpty(t, pkgs)
-			require.NotEmpty(t, relationships)
-			require.NoError(t, err)
-
-			if d := cmp.Diff(tt.wantRelationships, abstractRelationships(t, relationships)); d != "" {
-				t.Errorf("unexpected relationships (-want +got):\n%s", d)
-			}
 		})
 	}
 }
